feat: separate file for Destination.kt
